+++ lib/entao_dutil.dart
@@ -20,6 +20,7 @@ export 'src/ini_file.dart';
 export 'src/json_result.dart';
 export 'src/label_value.dart';
 export 'src/modbus.dart';
+export 'src/msg_call.dart';
 export 'src/msg_center.dart';
 export 'src/notify_value.dart';
 export 'src/nums.dart';