+++ test/map_ignore_case_test.dart
@@ -4,7 +4,7 @@ import 'package:test/test.dart';
 void main() {
   group('Ignore case map tests', () {
     test('Equal Test', () {
-      Map<String, int> map = mapIgnoreCase();
+      Map<String, int> map = ICaseMap();
       map["a"] = 11;
       map["A"] = 22;
       expect(22, map["a"]);